+++ CHANGELOG.md
@@ -23,6 +23,8 @@ The following changes have been completed, and will be included in the next tagg
 
 ### New
 
+* Basic Types
+  - added `validateNonEmptyArray()`
 * Errors
   - added `ArrayCannotBeEmptyError`